Moved Controller Generator and updated it to use the Generator Class
